Images of the change
Summary:
Here, I introduce change management and change analysis through this standard paper. Change is quite challenging to implement in the real world. In this paper, we will discuss how we can implement change. After analyzing various articles and books, the three fundamental change strategies must be followed by each human planning, training, and execution. Furthermore, the risk measurement before applying change is a crucial and challenging part of implementing change. This paper discusses how effectively we can apply the change and what precautions we need to measure. Also, in this paper most of the decisions made by the directors in the organization. Few of the decisions allowed by navigator and caretaker.